Added security checks for admin functions, some general cleanup

+++ remove_opp.php
@@ -1,21 +1,23 @@
 <?php
 include("db_connect.php");
 
+if ($_SESSION["role"] !== 'admin') {
+    header("location: access_denied.php");
+} else {
+    if($_SERVER["REQUEST_METHOD"] == "GET"){
+        
+        $opp_id = $_GET["id"];
 
-if($_SERVER["REQUEST_METHOD"] == "GET"){
-    
-    $opp_id = $_GET["id"];
+        $sql = "DELETE FROM opportunities WHERE opp_id = " . $opp_id;
+        if($conn->query($sql) === TRUE){
+            echo "Successfully deleted";
+        }
+        else{
+            echo "Error deleting resource: " . $conn->error;
+        }
 
-    $sql = "DELETE FROM opportunities WHERE opp_id = " . $opp_id;
-    if($conn->query($sql) === TRUE){
-        echo "Successfully deleted";
+        header("location:opportunities.php");
+
     }
-    else{
-        echo "Error deleting resource: " . $conn->error;
-    }
-
-    header("location:opportunities.php");
-
 }
-
 ?>
